+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+- Fix startup crash of Warhead launcher when Logitech G15 keyboard is connected
+([#39](https://github.com/ccomrade/c1-launcher/issues/39)).
+
 ## [v6] - 2024-02-15
 ### Added
 - Crysis Editor (Sandbox 2) support.